Cleanup entity generator.
